chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-same-url-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
FAIL target1.import should be non-null. Was undefined
PASS target1.import is target2.import
PASS target1.import is target3.import
PASS successfullyParsed is true

TEST COMPLETE
